updating build-image-azure.sh to understand how to build images for all clouds

+++ scripts/elotl/configure.sh
@@ -1,10 +1,16 @@
 #!/bin/sh
 
+# At this time, this script takes only 1 argument: what cloud we're running on
+CLOUD_PROVIDER=$1
+
 _step_counter=0
 step() {
 	_step_counter=$(( _step_counter + 1 ))
 	printf '\n\033[1;36m%d) %s\033[0m\n' $_step_counter "$@" >&2  # bold cyan
 }
+
+
+echo "Running configure.sh for cloud provider $CLOUD_PROVIDER"
 
 step "Update packages"
 apk update && apk upgrade
@@ -32,28 +38,23 @@ step 'Create ld-linux-x86-64.so.2 link'
 mkdir -p /lib64
 ln -s /lib/libc.musl-x86_64.so.1 /lib64/ld-linux-x86-64.so.2
 
-# todo: consider using
-# step "Setting root password to '$PW'"
-# PW="$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)"
-# echo -en "$PW\n$PW\n" | passwd root
-
 step 'Set password for root'
 echo 'root:*' | chpasswd -e
 
-step 'Install ssh authorized keys'
-mkdir -p /root/.ssh
-chmod 0700 /root/.ssh
-cat > /root/.ssh/authorized_keys <<-EOF
-ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQDPU7h8CaYA1VH/CwY3Ahw0s0wPbB/8t7A96GX/6qS2a8n79nGjywThZJP15L7NnTdTrdV59NEC1QvS0ym/JhlpokXwgMWURsYOAP5Y2lmK7wvAZ65bDn0iPiXXgyWPtEWQqCTV0U9HfZ81m+JMzfcED+L3w0iZAHSeRlupPZRtea3izx91A19RRn0NyVtmrwF4h3g537p+0O3DvaktxZddnwa3vPbY3CE6Eijiqsy9HOrx49YJS3SdBMvGNx91pynVLPWTCziBmYZCt8ioTGNvF8YWLVRf6VCj6M9zTG2NkCbXydAxpfRByTa+4yyKE44hmAehDM15pQGlmcg0O4HlepTqOvPZVyWAvkO3aD1xycrWSTKu68IgRzm9Ve064h3OUqVcWx1tybEAGyioC/H/vdJ4BGKH1wfQQvRbWrO8gCAr8LGS8JUIWWDPOCBtFobsyMo2opck9t8iM8lAiscueMNTJeRuIeK6692m0OsXL9+g8lHJkTD97VF963liCeRhaIG3kIaYXTyOhQdKbDQShT/r4yC8eMWDR/I6ab+2ir/qew46XwHJ98c/Ux0zII5v252D5Q/A4Wf6HJGOjAoMx4iQJ8Q5LYpLnIcX1WqznJQx1zPpaI9WpFe4ELK/mBAv53Emp3HjfacI74/RM6Zt/EHYcKV3Cr5VMaxgLaReaw== bcox@elotl.co
-ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQC/ywpZThNo+8665ftf8Se5hDfPewxsqvUiT2MifeBSHCHf0VSWKKDa9BG0o/gKVjEYsxrHBzR14SsPbhn2khP1jTKU1JPrSymqv2Zt5StGwadAWwXMn748zAu7mnIbioNn5pZ70K2eq/Q4LskQe0Sl9xpj92sPfLTEt3eAsPolIgvacy+qmlGmF7Zm+sFmmj+AvFyEFlhbsv8s92iwlGCNMXO2BkRTzYsTaLUVcf+xXbdxrxhTp15TnxzbMvBYTGSCnnCzK619zwDlY4MK33QWr3tEGVZalFXFX1NqRZbo1W8mfVHCG4bP/nMZZrROMTPmjl/KsMTC+4xRdzBlUO+7Fwd0qBEWyMe/90dtZajvXE20quLj7Dbbbx1UVuAXQXeiw5vTwPcwTBFqrbJ9dTyUqedQ2I/H8UeeQFkpTn3BC5yuccbpnVJg0yOBB4F4nUUdBSXCTTyz6I4EsBjhz12pyhGPljTgkcXJWmiv5bybD0EQQ+zZ4a077Ev9HTYo65xsCHr+Vh6KEy5TykV71lePgnfKpFTcsR5okuN9me8wRVpuytAZevnDwUhYblpBOl4GJo1yTZeQ+vPWKM3FJ7DvBY8w4TOdYsubcZwLcVIxEfMwTntohOOtATDQs8UlvLvhLPqtACDkpdHokb2YoDZFnewXwplReoe3hw7XlOb3cw== vilmos@x1
-ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCLoV/djnLOJt7gRprssm36m3kn2kwSYbjyh93A2/oHNcPM1rvONdeKQ2qAK/DbRVjGIIdcskDQWCAxTOsN4OEBkIOHsWlfE3TJ8f+IqGYyG1Ly+I66GIulWqZQfRb8H+4xKxVZ9C5g8eM58tYOAsToDFa/IX7RX2Ayu0BNc92+cUV8w92mL2c76MCwbKviipK7T6lDVn6IINxtJHVHwFZYonIiBYKn/30zGrL17sMbOVVd27lYPI/To61ui5PFSym7d/BOEkKRQKRgnHfpwm+VhXUuOGCv1FIeBZSs95JVhbIzynld8NBY3SjIGWi3m22HXR8yYlXRNZwB+BSrcWKL madhuri@elotl.co
-EOF
-chmod 0600 /root/.ssh/authorized_keys
-cat /root/.ssh/authorized_keys
+# step 'Install ssh authorized keys'
+# mkdir -p /root/.ssh
+# chmod 0700 /root/.ssh
+# cat > /root/.ssh/authorized_keys <<-EOF
+# ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQDPU7h8CaYA1VH/CwY3Ahw0s0wPbB/8t7A96GX/6qS2a8n79nGjywThZJP15L7NnTdTrdV59NEC1QvS0ym/JhlpokXwgMWURsYOAP5Y2lmK7wvAZ65bDn0iPiXXgyWPtEWQqCTV0U9HfZ81m+JMzfcED+L3w0iZAHSeRlupPZRtea3izx91A19RRn0NyVtmrwF4h3g537p+0O3DvaktxZddnwa3vPbY3CE6Eijiqsy9HOrx49YJS3SdBMvGNx91pynVLPWTCziBmYZCt8ioTGNvF8YWLVRf6VCj6M9zTG2NkCbXydAxpfRByTa+4yyKE44hmAehDM15pQGlmcg0O4HlepTqOvPZVyWAvkO3aD1xycrWSTKu68IgRzm9Ve064h3OUqVcWx1tybEAGyioC/H/vdJ4BGKH1wfQQvRbWrO8gCAr8LGS8JUIWWDPOCBtFobsyMo2opck9t8iM8lAiscueMNTJeRuIeK6692m0OsXL9+g8lHJkTD97VF963liCeRhaIG3kIaYXTyOhQdKbDQShT/r4yC8eMWDR/I6ab+2ir/qew46XwHJ98c/Ux0zII5v252D5Q/A4Wf6HJGOjAoMx4iQJ8Q5LYpLnIcX1WqznJQx1zPpaI9WpFe4ELK/mBAv53Emp3HjfacI74/RM6Zt/EHYcKV3Cr5VMaxgLaReaw== bcox@elotl.co
+# ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQC/ywpZThNo+8665ftf8Se5hDfPewxsqvUiT2MifeBSHCHf0VSWKKDa9BG0o/gKVjEYsxrHBzR14SsPbhn2khP1jTKU1JPrSymqv2Zt5StGwadAWwXMn748zAu7mnIbioNn5pZ70K2eq/Q4LskQe0Sl9xpj92sPfLTEt3eAsPolIgvacy+qmlGmF7Zm+sFmmj+AvFyEFlhbsv8s92iwlGCNMXO2BkRTzYsTaLUVcf+xXbdxrxhTp15TnxzbMvBYTGSCnnCzK619zwDlY4MK33QWr3tEGVZalFXFX1NqRZbo1W8mfVHCG4bP/nMZZrROMTPmjl/KsMTC+4xRdzBlUO+7Fwd0qBEWyMe/90dtZajvXE20quLj7Dbbbx1UVuAXQXeiw5vTwPcwTBFqrbJ9dTyUqedQ2I/H8UeeQFkpTn3BC5yuccbpnVJg0yOBB4F4nUUdBSXCTTyz6I4EsBjhz12pyhGPljTgkcXJWmiv5bybD0EQQ+zZ4a077Ev9HTYo65xsCHr+Vh6KEy5TykV71lePgnfKpFTcsR5okuN9me8wRVpuytAZevnDwUhYblpBOl4GJo1yTZeQ+vPWKM3FJ7DvBY8w4TOdYsubcZwLcVIxEfMwTntohOOtATDQs8UlvLvhLPqtACDkpdHokb2YoDZFnewXwplReoe3hw7XlOb3cw== vilmos@x1
+# ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCLoV/djnLOJt7gRprssm36m3kn2kwSYbjyh93A2/oHNcPM1rvONdeKQ2qAK/DbRVjGIIdcskDQWCAxTOsN4OEBkIOHsWlfE3TJ8f+IqGYyG1Ly+I66GIulWqZQfRb8H+4xKxVZ9C5g8eM58tYOAsToDFa/IX7RX2Ayu0BNc92+cUV8w92mL2c76MCwbKviipK7T6lDVn6IINxtJHVHwFZYonIiBYKn/30zGrL17sMbOVVd27lYPI/To61ui5PFSym7d/BOEkKRQKRgnHfpwm+VhXUuOGCv1FIeBZSs95JVhbIzynld8NBY3SjIGWi3m22HXR8yYlXRNZwB+BSrcWKL madhuri@elotl.co
+# EOF
+# chmod 0600 /root/.ssh/authorized_keys
+# cat /root/.ssh/authorized_keys
 
-step 'Enable ssh login for root'
-echo "PermitRootLogin yes" >> /etc/ssh/sshd_config
-cat /etc/ssh/sshd_config
+# step 'Enable ssh login for root'
+# echo "PermitRootLogin yes" >> /etc/ssh/sshd_config
+# cat /etc/ssh/sshd_config
 
 step 'Create itzo init script'
 cat > /etc/init.d/itzo <<-EOF
@@ -125,47 +126,42 @@ step 'Add cloud-init'
 wget -O /usr/local/bin/itzo-cloud-init http://itzo-dev-download.s3.amazonaws.com/itzo-cloud-init
 chmod 755 /usr/local/bin/itzo-cloud-init
 
+if [[ "$CLOUD_PROVIDER" == "aws" ]]; then
+    echo "http://dl-3.alpinelinux.org/alpine/edge/community" >> /etc/apk/repositories
+    apk update
+    apk add aws-ena-driver-virt
 
-echo "http://dl-3.alpinelinux.org/alpine/edge/community" >> /etc/apk/repositories
-apk update
-apk add aws-ena-driver-virt
+    # step 'Add aws-ena module'
+    # wget http://itzo-packages.s3.amazonaws.com/aws-ena-module.tar.gz
+    # tar xvzf aws-ena-module.tar.gz
+    # for kernel in /lib/modules/*; do
+    #     mkdir -p "${kernel}/kernel/drivers/net/ethernet/amazon/ena/"
+    #     cp ena.ko "${kernel}/kernel/drivers/net/ethernet/amazon/ena/"
+    #     depmod -a "$(basename ${kernel})"
+    # done
+    # rm aws-ena-module.tar.gz
+    echo ena > /etc/modules-load.d/ena.conf
 
-# step 'Add aws-ena module'
-# wget http://itzo-packages.s3.amazonaws.com/aws-ena-module.tar.gz
-# tar xvzf aws-ena-module.tar.gz
-# for kernel in /lib/modules/*; do
-#     mkdir -p "${kernel}/kernel/drivers/net/ethernet/amazon/ena/"
-#     cp ena.ko "${kernel}/kernel/drivers/net/ethernet/amazon/ena/"
-#     depmod -a "$(basename ${kernel})"
-# done
-# rm aws-ena-module.tar.gz
-echo ena > /etc/modules-load.d/ena.conf
+    # Taken from https://github.com/mcrute/alpine-ec2-ami/blob/master/make_ami.sh
+    # Create ENA feature for mkinitfs
+    # Submitted upstream: https://github.com/alpinelinux/mkinitfs/pull/19
+    echo "kernel/drivers/net/ethernet/amazon/ena" > /etc/mkinitfs/features.d/ena.modules
+    # Enable ENA and NVME features these don't hurt for any instance and are
+    # hard requirements of the 5 series and i3 series of instances
+    sed -Ei 's/^features="([^"]+)"/features="\1 nvme ena"/' /etc/mkinitfs/mkinitfs.conf
+    /sbin/mkinitfs $(basename $(find /lib/modules/* -maxdepth 0))
+fi
 
-# Taken from https://github.com/mcrute/alpine-ec2-ami/blob/master/make_ami.sh
-# Create ENA feature for mkinitfs
-# Submitted upstream: https://github.com/alpinelinux/mkinitfs/pull/19
-echo "kernel/drivers/net/ethernet/amazon/ena" > /etc/mkinitfs/features.d/ena.modules
-# Enable ENA and NVME features these don't hurt for any instance and are
-# hard requirements of the 5 series and i3 series of instances
-sed -Ei 's/^features="([^"]+)"/features="\1 nvme ena"/' /etc/mkinitfs/mkinitfs.conf
-/sbin/mkinitfs $(basename $(find /lib/modules/* -maxdepth 0))
-
-step "Setup Azure Linux Agent"
-wget -O ./waagent.tar.gz https://github.com/elotl/WALinuxAgent/archive/master.tar.gz && \
-tar xvzf ./waagent.tar.gz && \
-cd WALinuxAgent-master && \
-python setup.py install && \
-cd .. && \
-rm -rf WALinuxAgent-master waagent.tar.gz
-# step "Setup Azure Linux Agent"
-# wget https://s3.amazonaws.com/itzo-bcox-download/waagent_v2.2.34.tar.gz && \
-# tar xvzf waagent_v2.2.34.tar.gz && \
-# cd WALinuxAgent && \
-# python setup.py install && \
-# cd .. && \
-# rm -rf WALinuxAgent waagent_v2.2.34.tar.gz
-
-cat > /etc/init.d/waagent <<EOF
+if [[ "$CLOUD_PROVIDER" == "azure" ]]; then
+    step "Setup Azure Linux Agent"
+    wget -O ./waagent.tar.gz https://github.com/elotl/WALinuxAgent/archive/master.tar.gz && \
+	tar xvzf ./waagent.tar.gz && \
+	cd WALinuxAgent-master && \
+	python setup.py install && \
+	cd .. && \
+	rm -rf WALinuxAgent-master waagent.tar.gz
+    
+    cat > /etc/init.d/waagent <<EOF
 #!/sbin/openrc-run
 export PATH=/usr/local/sbin:$PATH
 start() {
@@ -174,9 +170,9 @@ start() {
         eend $? "Failed to start waagent"
 }
 EOF
-chmod +x /etc/init.d/waagent
+    chmod +x /etc/init.d/waagent
 
-cat > /etc/waagent.conf <<EOF
+    cat > /etc/waagent.conf <<EOF
 Provisioning.Enabled=y
 Extensions.Enabled=n
 Provisioning.UseCloudInit=n
@@ -204,6 +200,8 @@ CGroups.EnforceLimits=n
 CGroups.Excluded=customscript,runcommand
 AutoUpdate.Enabled=n
 EOF
+
+fi
 
 # #
 # # Note: the driver and libcuda client libraries need to be in sync, e.g. both
@@ -321,7 +319,9 @@ rc-update add resizeroot default
 rc-update add net.lo boot
 rc-update add termencoding boot
 rc-update add haveged boot
-rc-update add waagent default
-rc-update add hv_fcopy_daemon default
-rc-update add hv_kvp_daemon default
-rc-update add hv_vss_daemon default
+if [[ "$CLOUD_PROVIDER" == "azure" ]]; then
+    rc-update add waagent default
+    rc-update add hv_fcopy_daemon default
+    rc-update add hv_kvp_daemon default
+    rc-update add hv_vss_daemon default
+fi
